Provide Number of R Packages by Library

Provide the number of R package by library in a data.frame

Usage

lib_summary(by = .data$LibPath, sizes = FALSE)

Arguments

by

Selects the group_by variable

sizes

Should sizes of libraries be calculated. Default 'FALSE'. can be accessed through insert roxygen skeleton name of parameter and document what it does

Value

a data.frame of R packages by library

Examples

lib_summary()
#> # A tibble: 3 × 2
#>   LibPath                             n
#>   <chr>                           <int>
#> 1 /home/runner/work/_temp/Library    68
#> 2 /opt/R/4.4.2/lib/R/library         29
#> 3 /opt/R/4.4.2/lib/R/site-library     1